add todos, trlweRotateMinusOne8(done), trgswExternalProductSplitApproxCRT(todo), controlMuxApproxCRT(todo), trlgswExternalProduct(wip), trglevDotMultConst(debug), TrlweDotMultLargeConstants(wip), trgswExternalProductSplitNtt(wip), trlgsw(wip), RgswMultTestNTT14(wip), add second lvl decompose support for trgsw(wip), review todos, test decompose/ntt order (should have no diff), BlindRotLut(drift improve / wip) ntt mult/naive mult(small diff on large inputs)

+++ benchmark/crt.cpp
@@ -11,7 +11,7 @@ int main() {
     std::vector<int> lowModuli = {233, 239};
     std::vector<int> highModuli = {241, 251};
     long q = 3368562317;
-    std::vector<std::vector<int>> f(l, std::vector<int>(coeffs.size(), 0));
+    std::vector<std::vector<int8_t>> f(l, std::vector<int8_t>(coeffs.size(), 0));
     std::vector<int> f_tilde(coeffs.size(), 0);
     std::vector<long> w(l, 0l);
